helios-testing: Move deploy logic from TemporaryJob
to `DefaultDeployer`.

This is pretty much just shuffling code around. There's refactoring to do
to make things more reasonable.
