+++ helios-testing/src/main/java/com/spotify/helios/testing/TemporaryJob.java
@@ -17,74 +17,43 @@
 
 package com.spotify.helios.testing;
 
-import com.google.common.base.Optional;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Lists;
 import com.google.common.net.HostAndPort;
-import com.google.common.util.concurrent.Futures;
 
 import com.spotify.helios.client.HeliosClient;
-import com.spotify.helios.common.descriptors.Deployment;
-import com.spotify.helios.common.descriptors.Goal;
-import com.spotify.helios.common.descriptors.HostStatus;
 import com.spotify.helios.common.descriptors.Job;
-import com.spotify.helios.common.descriptors.JobStatus;
 import com.spotify.helios.common.descriptors.PortMapping;
 import com.spotify.helios.common.descriptors.TaskStatus;
-import com.spotify.helios.common.descriptors.ThrottleState;
-import com.spotify.helios.common.protocol.CreateJobResponse;
-import com.spotify.helios.common.protocol.JobDeployResponse;
-
-import org.apache.commons.lang.text.StrSubstitutor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
-import java.util.concurrent.Callable;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeoutException;
-import java.util.concurrent.atomic.AtomicBoolean;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
-import static com.google.common.base.Strings.isNullOrEmpty;
-import static com.google.common.collect.Maps.newHashMap;
-import static com.spotify.helios.testing.Jobs.TIMEOUT_MILLIS;
-import static com.spotify.helios.testing.Jobs.get;
 import static com.spotify.helios.testing.Jobs.getJobDescription;
 import static java.lang.String.format;
-import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static org.junit.Assert.fail;
 
 public class TemporaryJob {
 
-  private static final Logger log = LoggerFactory.getLogger(TemporaryJob.class);
-
-  private final Map<String, TaskStatus> statuses = newHashMap();
+  private final Map<String, TaskStatus> statuses;
   private final HeliosClient client;
-  private final Prober prober;
   private final Job job;
   private final List<String> hosts;
-  private final Map<String, String> hostToIp = newHashMap();
-  private final Set<String> waitPorts;
-  private final String jobDeployedMessageFormat;
-  private final long deployTimeoutMillis;
+  private final Map<String, String> hostToIp;
 
-  TemporaryJob(final HeliosClient client, final Prober prober,
+  TemporaryJob(final HeliosClient client,
                final Job job,
-               final List<String> hosts, final Set<String> waitPorts,
-               final String jobDeployedMessageFormat, final long deployTimeoutMillis) {
+               final List<String> hosts,
+               final Map<String, String> hostToIp,
+               final Map<String, TaskStatus> statuses) {
     this.client = checkNotNull(client, "client");
-    this.prober = checkNotNull(prober, "prober");
     this.job = checkNotNull(job, "job");
     this.hosts = ImmutableList.copyOf(checkNotNull(hosts, "hosts"));
-    this.waitPorts = ImmutableSet.copyOf(checkNotNull(waitPorts, "waitPorts"));
-    this.jobDeployedMessageFormat = Optional.fromNullable(jobDeployedMessageFormat).or("");
-    this.deployTimeoutMillis = deployTimeoutMillis;
+    this.hostToIp = ImmutableMap.copyOf(checkNotNull(hostToIp, "hostToIp"));
+    this.statuses = ImmutableMap.copyOf(checkNotNull(statuses, "statuses"));
   }
 
   public Job job() {
@@ -96,7 +65,7 @@ public class TemporaryJob {
   }
 
   public Map<String, TaskStatus> statuses() {
-    return ImmutableMap.copyOf(statuses);
+    return statuses;
   }
 
   /**
@@ -162,54 +131,6 @@ public class TemporaryJob {
     return addresses;
   }
 
-  void deploy() {
-    try {
-      // Create job
-      log.info("Creating job {}", job.getId().toShortString());
-      final CreateJobResponse createResponse = get(client.createJob(job));
-      if (createResponse.getStatus() != CreateJobResponse.Status.OK) {
-        fail(format("Failed to create job %s - %s", job.getId(),
-                    createResponse.toString()));
-      }
-    } catch (InterruptedException | ExecutionException | TimeoutException e) {
-      fail(format("Failed to create job %s %s - %s", job.getId(), job.toString(), e));
-    }
-
-    try {
-      // Deploy job
-      final Deployment deployment = Deployment.of(job.getId(), Goal.START);
-      for (final String host : hosts) {
-        // HELIOS_HOST_ADDRESS is the IP address we should use to reach the host, instead of
-        // the hostname. This is used when running a helios cluster inside a VM, and the containers
-        // can be reached by IP address only, since DNS won't be able to resolve the host name of
-        // the helios agent running in the VM.
-        final HostStatus hostStatus = client.hostStatus(host).get();
-        final String hostAddress = hostStatus.getEnvironment().get("HELIOS_HOST_ADDRESS");
-        if (hostAddress != null) {
-          hostToIp.put(host, hostAddress);
-        }
-
-        log.info("Deploying {} to {}", getJobDescription(job), host);
-        final JobDeployResponse deployResponse = get(client.deploy(deployment, host));
-        if (deployResponse.getStatus() != JobDeployResponse.Status.OK) {
-          fail(format("Failed to deploy job %s %s - %s",
-                      job.getId(), job.toString(), deployResponse));
-        }
-      }
-    } catch (InterruptedException | ExecutionException | TimeoutException e) {
-      fail(format("Failed to deploy job %s %s - %s", job.getId(), job.toString(), e));
-    }
-
-    try {
-      // Wait for job to come up
-      for (final String host : hosts) {
-        awaitUp(host);
-      }
-    } catch (TimeoutException e) {
-      fail(format("Failed while probing job %s %s - %s", job.getId(), job.toString(), e));
-    }
-  }
-
   void undeploy(final List<AssertionError> errors) {
     Jobs.undeploy(client, job, hosts, errors);
   }
@@ -228,104 +149,6 @@ public class TemporaryJob {
                   getJobDescription(job), errors.get(0)));
     }
   }
-  
-  private void awaitUp(final String host) throws TimeoutException {
-    final AtomicBoolean messagePrinted = new AtomicBoolean(false);
-    final TaskStatus status = Polling.awaitUnchecked(
-      deployTimeoutMillis, MILLISECONDS, new Callable<TaskStatus>() {
-        @Override
-        public TaskStatus call() throws Exception {
-          final JobStatus status = Futures.getUnchecked(client.jobStatus(job.getId()));
-          if (status == null) {
-            log.debug("Job status not available");
-            return null;
-          }
-          final TaskStatus taskStatus = status.getTaskStatuses().get(host);
-          if (taskStatus == null) {
-            log.debug("Task status not available on {}", host);
-            return null;
-          }
-
-          if (!messagePrinted.get() &&
-              !isNullOrEmpty(jobDeployedMessageFormat) &&
-              !isNullOrEmpty(taskStatus.getContainerId())) {
-            outputDeployedMessage(host, taskStatus.getContainerId());
-            messagePrinted.set(true);
-          }
-
-          verifyHealthy(host, taskStatus);
-
-          final TaskStatus.State state = taskStatus.getState();
-          log.info("Job state of {}: {}", job.getImage(), state);
-
-          if (state == TaskStatus.State.RUNNING) {
-            return taskStatus;
-          }
-
-          return null;
-        }
-      }
-    );
-
-    statuses.put(host, status);
-
-    for (final String port : waitPorts) {
-      awaitPort(port, host);
-    }
-  }
-
-  void verifyHealthy() throws AssertionError {
-    log.debug("Checking health of {}", job.getImage());
-    final JobStatus status = Futures.getUnchecked(client.jobStatus(job.getId()));
-    if (status == null) {
-      return;
-    }
-    for (final Map.Entry<String, TaskStatus> entry : status.getTaskStatuses().entrySet()) {
-      verifyHealthy(entry.getKey(), entry.getValue());
-    }
-  }
-
-  private void verifyHealthy(final String host, final TaskStatus status) {
-    log.debug("Checking health of {} on {}", job.getImage(), host);
-    final TaskStatus.State state = status.getState();
-    if (state == TaskStatus.State.FAILED ||
-        state == TaskStatus.State.EXITED ||
-        state == TaskStatus.State.STOPPED) {
-      // Throw exception which should stop the test dead in it's tracks
-      String stateString = state.toString();
-      if (status.getThrottled() != ThrottleState.NO) {
-        stateString += format("(%s)", status.getThrottled());
-      }
-      throw new AssertionError(format(
-          "Unexpected job state %s for job %s with image %s on host %s. Check helios agent "
-          + "logs for details. If you're using HeliosSoloDeployment, set "
-          + "`HeliosSoloDeployment.fromEnv().removeHeliosSoloOnExit(false)` and check the"
-          + "logs of the helios-solo container with `docker logs <container ID>`.",
-          stateString, job.getId().toShortString(), job.getImage(), host));
-    }
-  }
-
-  private void awaitPort(final String port, final String host) throws TimeoutException {
-    final String endpoint = endpointFromHost(host);
-    final TaskStatus taskStatus = statuses.get(host);
-    assert taskStatus != null;
-    final PortMapping portMapping = taskStatus.getPorts().get(port);
-    final Integer externalPort = portMapping.getExternalPort();
-    assert externalPort != null;
-    Polling.awaitUnchecked(TIMEOUT_MILLIS, MILLISECONDS, new Callable<Boolean>() {
-      @Override
-      public Boolean call() throws Exception {
-        log.info("Probing: {} @ {}:{}", port, endpoint, portMapping);
-        final boolean up = prober.probe(endpoint, portMapping);
-        if (up) {
-          log.info("Up: {} @ {}:{}", port, endpoint, externalPort);
-          return true;
-        } else {
-          return null;
-        }
-      }
-    });
-  }
 
   /**
    * Returns the ip address mapped to the given hostname. If no mapping exists, the hostname is
@@ -336,18 +159,5 @@ public class TemporaryJob {
   private String endpointFromHost(String host) {
     final String ip = hostToIp.get(host);
     return ip == null ? host : ip;
-  }
-
-  private void outputDeployedMessage(final String host, final String containerId) {
-    final StrSubstitutor subst = new StrSubstitutor(new ImmutableMap.Builder<String, Object>()
-        .put("host", host)
-        .put("name", job.getId().getName())
-        .put("version", job.getId().getVersion())
-        .put("hash", job.getId().getHash())
-        .put("job", job.toString())
-        .put("containerId", containerId)
-        .build()
-        );
-    log.info("{}", subst.replace(jobDeployedMessageFormat));
   }
 }
